+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Date format: DD/MM/YYYY
 
 ## [3.4.2] - [DD/MM/YYYY]
 
+- `NavigationPane` width can now be customizable ([#99](https://github.com/bdlukaa/fluent_ui/issues/99)) 
 - Implement `PaneItemAction` for `NavigationPane` ([#104](https://github.com/bdlukaa/fluent_ui/issues/104))
 
 ## [3.4.1] - [08/11/2021]